mcp/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp[cli]==1.21.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
packaging==25.0
pinecone-client==3.0.0
praw==7.7.1
prawcore==2.4.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
safetensors==0.6.2
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
tokenizers==0.15.2
torch==2.2.0
tqdm==4.66.1
transformers==4.36.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.38.0
websocket-client==1.9.0
